migration fixed
Former-commit-id: 5847d3edbf62beb7f875941a7e847199d92561d1

+++ database/migrations/2018_10_09_165251_modify_columns_in_files_table.php
@@ -14,20 +14,12 @@ class ModifyColumnsInFilesTable extends Migration
     public function up()
     {
         //this migration removes start_date, end_date, recommendation, rejection_reason from files table
-        Schema::table('files', function (Blueprint $table) {
-            $table->dropColumn(['start_date', 'end_date', 'recommendation', 'rejection_reason', 'is_file_accepted', 'status']);
-        });
+        if(Schema::hasColumn('start_date', 'end_date', 'recommendation', 'rejection_reason', 'is_file_accepted', 'status')){
+            Schema::table('files', function (Blueprint $table) {
+                $table->dropColumn(['start_date', 'end_date', 'recommendation', 'rejection_reason', 'is_file_accepted', 'status']);
+            });
+        }
+
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::table('file', function (Blueprint $table) {
-            //
-        });
-    }
 }
